Properly handle specified test names
Before, the available tests were just checked against the passed in
comma-separated string of test names, and if they were found in the string they
were run. This causes problems when you have two tests like "test" and
"bettertest", because if you want to run "bettertest", it runs "test" as well.
This fixes the problem by splitting the input on commas (it specifies that the
list of tests should be comma-separated in the usage), so then the `in` test
below checks the right thing.